+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -10,7 +10,6 @@ public class ExperimentalRuleSetProvider : RuleSetProvider {
         experimentalRulesetId,
         MultiLineIfElseRule(),
         NoEmptyFirstLineInMethodBlockRule(),
-        SpacingBetweenDeclarationsWithCommentsRule(),
         UnnecessaryParenthesesBeforeTrailingLambdaRule(),
         TypeParameterListSpacingRule(),
         TypeArgumentListSpacingRule(),