+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/SpacingBetweenDeclarationsWithCommentsRule.kt
@@ -1,4 +1,4 @@
-package com.pinterest.ktlint.ruleset.experimental
+package com.pinterest.ktlint.ruleset.standard
 
 import com.pinterest.ktlint.core.Rule
 import com.pinterest.ktlint.core.ast.ElementType.FILE
@@ -16,7 +16,7 @@ import org.jetbrains.kotlin.psi.psiUtil.startOffset
 /**
  * @see https://youtrack.jetbrains.com/issue/KT-35088
  */
-class SpacingBetweenDeclarationsWithCommentsRule : Rule("$experimentalRulesetId:spacing-between-declarations-with-comments") {
+class SpacingBetweenDeclarationsWithCommentsRule : Rule("spacing-between-declarations-with-comments") {
     override fun visit(
         node: ASTNode,
         autoCorrect: Boolean,