+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -47,6 +47,7 @@ public class StandardRuleSetProvider : RuleSetProvider {
         SpacingAroundRangeOperatorRule(),
         SpacingAroundUnaryOperatorRule(),
         SpacingBetweenDeclarationsWithAnnotationsRule(),
+        SpacingBetweenDeclarationsWithCommentsRule(),
         StringTemplateRule(),
         TrailingCommaRule(),
         WrappingRule()