+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The rules below are promoted from the Experiment ruleset to the Standard ruleset
 * `traling-comma`
 * `spacing-around-angle-brackets`
 * `spacing-between-declarations-with-annotations`
+* `spacing-between-declarations-with-comments`
 * `unary-op-spacing`
 
 Note that as a result of moving the rules that the prefix `experimental:` has to be removed from all references to this rule. Check references in: